requirements.txt
numpy==1.24.3
streamlit==1.35.0
pillow==9.4.0
tensorflow==2.16.1
opencv-python-headless==4.10.0.82
